+++ src/main/java/insane96mcp/enhancedai/modules/villager/feature/VillagerAttacking.java
@@ -1,6 +1,7 @@
 package insane96mcp.enhancedai.modules.villager.feature;
 
 import insane96mcp.enhancedai.modules.Modules;
+import insane96mcp.enhancedai.modules.villager.EAVillagerHurtByTargetGoal;
 import insane96mcp.enhancedai.setup.EAStrings;
 import insane96mcp.enhancedai.setup.NBTUtils;
 import insane96mcp.insanelib.base.Feature;
@@ -11,7 +12,6 @@ import insane96mcp.insanelib.base.config.Config;
 import insane96mcp.insanelib.base.config.LoadFeature;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.world.entity.ai.goal.MeleeAttackGoal;
-import net.minecraft.world.entity.ai.goal.target.HurtByTargetGoal;
 import net.minecraft.world.entity.npc.Villager;
 import net.minecraftforge.event.entity.EntityJoinLevelEvent;
 import net.minecraftforge.eventbus.api.EventPriority;
@@ -22,10 +22,15 @@ import java.util.List;
 @Label(name = "Villager Attacking", description = "Make villagers fight back")
 @LoadFeature(module = Modules.Ids.VILLAGER)
 public class VillagerAttacking extends Feature {
-    //TODO add attack back on low reputation only
     @Config
     @Label(name = "Villagers Fight back", description = "If true, when attacked, villagers will call other villagers for help and attack back. Attack damage can't be changed via config due to limitation so use mods like Mobs Properties Randomness to change the damage. Base damage is 4")
     public static Boolean villagersFightBack = true;
+    @Config
+    @Label(name = "Villagers Fight back Enemies", description = "If false villagers will not attack back monsters")
+    public static Boolean villagersFightBackEnemies = false;
+    @Config
+    @Label(name = "Reputation for Fight back", description = "Villagers will only attack players that have below this reputation (like Iron Golems by default). https://minecraft.fandom.com/wiki/Villager#Gossiping")
+    public static Integer minReputationFightBack = -100;
     @Config(min = 0d, max = 4d)
     @Label(name = "Movement Speed Multiplier", description = "Movement speed multiplier when attacking")
     public static Double speedMultiplier = 0.8d;
@@ -50,7 +55,7 @@ public class VillagerAttacking extends Feature {
         double movementSpeedMultiplier = NBTUtils.getDoubleOrPutDefault(persistentData, EAStrings.Tags.Passive.SPEED_MULTIPLIER_WHEN_AGGROED, speedMultiplier);
 
         if (villagersFightBack) {
-            villager.targetSelector.addGoal(1, (new HurtByTargetGoal(villager)).setAlertOthers());
+            villager.targetSelector.addGoal(1, (new EAVillagerHurtByTargetGoal(villager)).setAlertOthers());
             villager.goalSelector.addGoal(1, new MeleeAttackGoal(villager, movementSpeedMultiplier, false));
             //villager.getBrain().removeAllBehaviors();
         }